msm-poweroff: Initialize display configuration in imem for recovery
Set the initial value of the display configuration in imem to ensure proper
behavior during recovery scenarios, such as when the device encounters a panic.

By setting the first 4 bytes of the display configuration imem to `0xdead4ead`,
Whenever a panic occurs, the XBL ramdump displays an orange screen, with the
framebuffer address derived from these bytes.

Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Change-Id: Ifa38d1e13ca988d32f5efefe18c630c73f130711
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>